make the sidebar a component and change icons
